+++ discopy/compact.py
@@ -137,6 +137,17 @@ class Swap(Permutation, symmetric.Swap, ribbon.Braid, Box):
     """
 
 
+class Sum(symmetric.Sum, ribbon.Sum, Box):
+    """
+    A compact sum is a sum of compact diagrams.
+
+    Parameters:
+        terms (tuple[Diagram, ...]) : The terms of the formal sum.
+        dom (Ty) : The domain of the formal sum.
+        cod (Ty) : The codomain of the formal sum.
+    """
+
+
 class Functor(symmetric.Functor, ribbon.Functor):
     """
     A compact functor is both a symmetric functor and a ribbon functor.
@@ -164,6 +175,7 @@ Diagram.functor_factory = Functor
 Diagram.permutation_factory = Permutation
 Hypergraph = hypergraph.Hypergraph[Diagram]
 Diagram.cup_factory, Diagram.cap_factory = Cup, Cap
+Diagram.sum_factory = Sum
 
 
 class Equation(symmetric.Equation):